requirements.txt
Django>=4.2,<5
djangorestframework
psycopg2
